Factored out makeflow_summary.
